+++ src/GoldyBot/assets.py
@@ -1,3 +1,6 @@
+ATTACHMENT_URL = "attachment://"
+""""""
+
 NO_PFP_IMAGE = "https://media.discordapp.net/attachments/436201641486581762/967863279575765042/logo_v2.png"
 """"""