+++ src/GoldyBot/utility/commands/_send_modal_.py
@@ -1,0 +1,4 @@
+import GoldyBot
+
+async def send_modal(ctx:GoldyBot.InteractionToCtx, modal:GoldyBot.nextcord.ui.Modal) -> GoldyBot.objects.slash.Message:
+    return await ctx.send_modal(modal)