+++ src/GoldyBot/utility/views/__init__.py
@@ -1,1 +1,1 @@
-from . import confirm
+from . import confirm, dropdown, forms